Remove AWSClient EventLoop APIs (#553)
* Remove AWSClient.execute calls with EventLoop

* Fix STSAssumeRole after removing EventLoop execute

* Fix AWSClient waiter code

* Remove commented out code

* Remove EL credential functions

* Merge async files with parent implementations

* Delete AWSClient+Waiter+async.swift

* Fix up paginate code

* Remove EventLoop from execute closure

* Changes requested in PR

* Update Sources/SotoCore/Waiters/AWSClient+Waiter.swift


---------
