+++ Sources/SotoCore/AWSService.swift
@@ -42,44 +42,6 @@ extension AWSService {
     /// The EventLoopGroup service is using
     public var eventLoopGroup: EventLoopGroup { return client.eventLoopGroup }
 
-    /// Generate a signed URL
-    /// - parameters:
-    ///     - url : URL to sign
-    ///     - httpMethod: HTTP method to use (.GET, .PUT, .PUSH etc)
-    ///     - headers: Headers that are to be used with this URL. Be sure to include these headers when you used the returned URL
-    ///     - expires: How long before the signed URL expires
-    ///     - logger: Logger to output to
-    /// - returns:
-    ///     A signed URL
-    public func signURL(
-        url: URL,
-        httpMethod: HTTPMethod,
-        headers: HTTPHeaders = HTTPHeaders(),
-        expires: TimeAmount,
-        logger: Logger = AWSClient.loggingDisabled
-    ) -> EventLoopFuture<URL> {
-        return self.client.signURL(url: url, httpMethod: httpMethod, headers: headers, expires: expires, serviceConfig: self.config, logger: logger)
-    }
-
-    /// Generate signed headers
-    /// - parameters:
-    ///     - url : URL to sign
-    ///     - httpMethod: HTTP method to use (.GET, .PUT, .PUSH etc)
-    ///     - headers: Headers that are to be used with this URL. Be sure to include these headers when you used the returned URL
-    ///     - body: body payload to sign as well. While it is unnecessary to provide the body for S3 other services require it
-    ///     - logger: Logger to output to
-    /// - returns:
-    ///     A series of signed headers including the original headers provided to the function
-    public func signHeaders(
-        url: URL,
-        httpMethod: HTTPMethod,
-        headers: HTTPHeaders = HTTPHeaders(),
-        body: AWSPayload = .empty,
-        logger: Logger = AWSClient.loggingDisabled
-    ) -> EventLoopFuture<HTTPHeaders> {
-        return self.client.signHeaders(url: url, httpMethod: httpMethod, headers: headers, body: body, serviceConfig: self.config, logger: logger)
-    }
-
     /// Return new version of Service with edited parameters
     /// - Parameters:
     ///   - region: Server region
@@ -102,5 +64,43 @@ extension AWSService {
             byteBufferAllocator: byteBufferAllocator,
             options: options
         ))
+    }
+
+    /// Generate a signed URL
+    /// - parameters:
+    ///     - url : URL to sign
+    ///     - httpMethod: HTTP method to use (.GET, .PUT, .PUSH etc)
+    ///     - headers: Headers that are to be used with this URL. Be sure to include these headers when you used the returned URL
+    ///     - expires: How long before the signed URL expires
+    ///     - logger: Logger to output to
+    /// - returns:
+    ///     A signed URL
+    public func signURL(
+        url: URL,
+        httpMethod: HTTPMethod,
+        headers: HTTPHeaders = HTTPHeaders(),
+        expires: TimeAmount,
+        logger: Logger = AWSClient.loggingDisabled
+    ) async throws -> URL {
+        return try await self.client.signURL(url: url, httpMethod: httpMethod, headers: headers, expires: expires, serviceConfig: self.config, logger: logger)
+    }
+
+    /// Generate signed headers
+    /// - parameters:
+    ///     - url : URL to sign
+    ///     - httpMethod: HTTP method to use (.GET, .PUT, .PUSH etc)
+    ///     - headers: Headers that are to be used with this URL. Be sure to include these headers when you used the returned URL
+    ///     - body: body payload to sign as well. While it is unnecessary to provide the body for S3 other services require it
+    ///     - logger: Logger to output to
+    /// - returns:
+    ///     A series of signed headers including the original headers provided to the function
+    public func signHeaders(
+        url: URL,
+        httpMethod: HTTPMethod,
+        headers: HTTPHeaders = HTTPHeaders(),
+        body: AWSPayload = .empty,
+        logger: Logger = AWSClient.loggingDisabled
+    ) async throws -> HTTPHeaders {
+        return try await self.client.signHeaders(url: url, httpMethod: httpMethod, headers: headers, body: body, serviceConfig: self.config, logger: logger)
     }
 }
