+++ Sources/SotoCore/Waiters/AWSClient+Waiter.swift
@@ -39,7 +39,7 @@ extension AWSClient {
             let matcher: AWSWaiterMatcher
         }
 
-        public typealias WaiterCommand = @Sendable (Input, Logger, EventLoop?) -> EventLoopFuture<Output>
+        public typealias WaiterCommand = @Sendable (Input, Logger) async throws -> Output
 
         /// Initialize an waiter
         /// - Parameters:
@@ -89,6 +89,70 @@ extension AWSClient {
                 return remainingTime - self.minDelayTime
             }
             return timeDelay
+        }
+    }
+
+    /// Returns when waiter polling returns a success state
+    /// or returns an error if the polling returns an error or timesout
+    ///
+    /// - Parameters:
+    ///   - input: Input parameters
+    ///   - waiter: Waiter to wait on
+    ///   - maxWaitTime: Maximum amount of time to wait
+    ///   - logger: Logger used to provide output
+    public func waitUntil<Input, Output>(
+        _ input: Input,
+        waiter: Waiter<Input, Output>,
+        maxWaitTime: TimeAmount? = nil,
+        logger: Logger = AWSClient.loggingDisabled
+    ) async throws {
+        let maxWaitTime = maxWaitTime ?? waiter.maxDelayTime
+        let deadline: NIODeadline = .now() + maxWaitTime
+
+        var attempt = 0
+        while true {
+            attempt += 1
+            let result: Result<Output, Error>
+            do {
+                result = try .success(await waiter.command(input, logger))
+            } catch {
+                result = .failure(error)
+            }
+            var acceptorState: WaiterState?
+            for acceptor in waiter.acceptors {
+                if acceptor.matcher.match(result: result.map { $0 }) {
+                    acceptorState = acceptor.state
+                    break
+                }
+            }
+            // if state has not been set then set it based on return of API call
+            let waiterState: WaiterState
+            if let state = acceptorState {
+                waiterState = state
+            } else if case .failure = result {
+                waiterState = .failure
+            } else {
+                waiterState = .retry
+            }
+            // based on state succeed, fail promise or retry
+            switch waiterState {
+            case .success:
+                return
+            case .failure:
+                if case .failure(let error) = result {
+                    throw error
+                } else {
+                    throw ClientError.waiterFailed
+                }
+            case .retry:
+                let wait = waiter.calculateRetryWaitTime(attempt: attempt, remainingTime: deadline - .now())
+                if wait < .seconds(0) {
+                    throw ClientError.waiterTimeout
+                } else {
+                    logger.trace("Wait \(wait.nanoseconds / 1_000_000)ms")
+                    try await Task.sleep(nanoseconds: UInt64(wait.nanoseconds))
+                }
+            }
         }
     }
 }
